Move all department specific components into the proper module
