+++ app/components/admin/departments/detail_component.html.erb
