+++ app/components/admin/departments/organization_name_component.html.erb
