+++ app/views/admin/departments/index.html.erb
@@ -1,3 +1,3 @@
 <%= render(Admin::Departments::PageHeaderComponent.new) %>
 
-<%= render(Admin::Groups::GroupHierarchyLayoutComponent.new(groups: @groups, active_group: @active_group)) %>
+<%= render(Admin::Departments::HierarchyLayoutComponent.new(groups: @groups, active_group: @active_group)) %>
